+++ DaddysPlaceApi/ViewEntity/OrderViewEntity.cs
@@ -3,8 +3,8 @@
     public class OrderViewEntity
     {
         public int Id { get; set; }
-        public float Price { get; set; }
-        public float? Discount { get; set; }
-        public float Total { get; set; }
+        public decimal Price { get; set; }
+        public decimal? Discount { get; set; }
+        public decimal Total { get; set; }
     }
 }
